Episode 24

+++ app/Http/Controllers/ThreadsController.php
@@ -84,8 +84,17 @@ class ThreadsController extends Controller
         ]);
     }
 
+    /**
+     * Delete the given thread.
+     *
+     * @param        $channel
+     * @param Thread $thread
+     * @return mixed
+     */
     public function destroy($channel, Thread $thread)
     {
+        $this->authorize('update', $thread);
+
         $thread->delete();
 
         if (request()->wantsJson()) {
